setting up endpoints

+++ pages/api/endpointDelete.js
@@ -3,10 +3,10 @@ const db = require('../../lib/db')
 export default async function (req, res){
     try{
         console.log('req.body: ', req.body);
-        const query = `DELETE FROM endpointhistory(endpoint, userid)
-        VALUES ($1, $2)`
+        const query = `DELETE FROM endpointhistory(endpoint, userid) WHERE userid = ${1}`
+        //come bacl to this to get right query
         const values = [req.body.endpoint, req.body.userid]
-        const result = await conn.query(
+        const result = await db.query(
             query,
             values
         );
